Introduce CommonFunctions trait to house common methods/functions for all tasks to use.
Apply standard of task properties via the 'invokeOnNonNullOrException' method. This method will check for the existence of
a given method and if it exists will execute with possible args otherwise we thrown a meaningful exception. This was done
as the 'kubernetes-client' object has builders/fluents for everything and often when invoking a method you get a new
object returned. As such we'd like a clean way to guard against developers potentially setting properties on the tasks
themselves but also through the 'config{}' construct.